+++ demo/components/community/index.html
@@ -1,11 +1,12 @@
 <!DOCTYPE html>
 <html lang="en">
 	<head>
+		<title>Community Demo</title>
 		<link rel="stylesheet" href="/node_modules/@brightspace-ui/core/components/demo/styles.css" type="text/css">
 		<script type="module">
 			import '@brightspace-ui/core/components/demo/demo-page.js';
-			import '../../../src/components/community/link-community.js';
-			import '../../../src/components/community/button-community.js';
+			import '../../../src/components/community/community-link.js';
+			import '../../../src/components/community/community-button.js';
 		</script>
 		<meta name="viewport" content="width=device-width, initial-scale=1.0">
 		<meta charset="UTF-8">
