+++ index.html
@@ -21,7 +21,7 @@
 
 		<h2 class="d2l-heading-2">Components</h2>
 		<ul>
-			<li><a href="demo/components/community-link/index.html">Community</a></li>
+			<li><a href="demo/components/community/index.html">Community</a></li>
 			<li><a href="demo/components/opt-in-flyout/index.html">Opt-In Flyout</a></li>
 		</ul>
